+++ secrets/ed25519_test.go
@@ -7,4 +7,5 @@ func TestEd25519(t *testing.T) {
 	// TODO: LoadEd25519PrivateKey
 	// TODO: LoadEd25519PublicKey
 	// TODO: Ed25519SignMessage
+	// TODO: Ed25519VerifySignature
 }